Inline use-agent-initialization.pure.ts into use-agent-initialization.ts
